ditching lib folder and using config and utl folder instead
