+++ components/Header/index.tsx
@@ -1,4 +1,5 @@
-import { cn } from "@/lib/util";
+// UTILS
+import { cn } from '@/utils/helpers';
 
 interface HeaderProps extends React.HTMLAttributes<HTMLDivElement> {
   headerTitle?: React.ReactNode;
