+++ config/pricing.ts
